ta120_adapter: Refactor serializers to separate module
