+++ backend/ta120_adapter/endpoint.py
@@ -1,7 +1,7 @@
 import json
 from typing import Any, Callable, Mapping, Tuple
 
-from rest_framework import generics, permissions, serializers
+from rest_framework import generics, permissions
 from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.renderers import BrowsableAPIRenderer, JSONRenderer
 from rest_framework.response import Response
@@ -14,49 +14,22 @@ from datahubhel.gateway_utils import (
 
 from .authentication import SensorKeyAuthentication
 from .models import Sensor
+from .properties import DATA_PROPERTIES
+from .serializers import QueryParametersSerializer, SensorDataSerializer
 from .ul20 import UltraLight20Parser, UltraLight20Renderer
 from .utils import dump_laeq1s_registers
 
 TOPIC = 'fi.fvh.observations.noise.ta120'
 
-
-class QueryParametersSerializer(serializers.Serializer):
-    t = serializers.DateTimeField(
-        label="time", help_text="timestamp", required=False)
-    getCmd = serializers.BooleanField(  # noqa: N815
-        default=True, initial=True, required=False,
-        label="get_command", help_text="request for receive commands")
-
-
-class SensorDataSerializer(serializers.Serializer):
-    n = serializers.FloatField(
-        label="level", help_text="sound pressure level")
-    o = serializers.NullBooleanField(
-        label="overload", help_text="overload", required=False)
-    u = serializers.NullBooleanField(
-        label="underrange", help_text="underrange", required=False)
-    b = serializers.FloatField(
-        label="battery", help_text="battery level", required=False)
-    p = serializers.NullBooleanField(
-        label="power", help_text="power supply status", required=False)
-    w = serializers.FloatField(
-        label="wifi_strength", help_text="wi-fi strength", required=False)
-    m = serializers.FloatField(
-        label="modem_strength", help_text="modem strength", required=False)
-    s = serializers.CharField(
-        label="laeq1s_registers", help_text="LAeq1s registers", required=False)
-
-
 SERIALIZERS = {
     'laeq1s_registers': dump_laeq1s_registers,
 }
 
-
 # Map each property key to a pair of label and serializer,
 # e.g. "n" is mapped to ("level", json.dumps).
 PROPERTIES: Mapping[str, Tuple[str, Callable[[Any], str]]] = {
-    k: (v.label, SERIALIZERS.get(v.label, json.dumps))  # type: ignore
-    for (k, v) in SensorDataSerializer().fields.items()
+    key: (prop.label, SERIALIZERS.get(prop.label, json.dumps))
+    for (key, prop) in DATA_PROPERTIES.items()
 }
